【追加】【変更】リザルト画面レイアウト変更＆評価ランク機能実装

+++ RhythmGame/Assets/Scripts/GameScene/ScoreBoxWindow.cs
@@ -35,7 +35,7 @@ public class ScoreBoxWindow : MonoBehaviour
 
     public void UpdateScoreBoxWindow(ScoreData score)
     {
-        UpdateTxtScore(score.valueScore);
+        UpdateTxtScore((int)score.valueScore);
         UpdateTxtMaxCombo(score.valueMaxCombo);
         UpdateTxtPerfect(score.valuePerfect);
         UpdateTxtGreat(score.valueGreat);
